+++ example/extension_hats/hat_pir_motion.dart
@@ -7,33 +7,32 @@ import 'dart:io';
 
 import 'parse_cmd_line.dart';
 
-const wait = 150;
+const wait = 200;
 
-/// https://wiki.seeedstudio.com/Grove-Vibration_Sensor_SW-420/
-
+/// https://wiki.seeedstudio.com/Grove-PIR_Motion_Sensor/
 ///
-/// Usage: [nano|grove|grovePlus] vibrationPin ledPin
+/// Usage: [nano|grove|grovePlus] pirMotionPin ledPin
 void main(List<String> args) {
-  String pinInfo = "Vibration pin";
-  var tupple = checkArgs2Pins(args, "vibrationPin", "ledPin");
-  var vibrationPin = tupple.$2;
+  String pinInfo = "Magenetic switch pin";
+  var tupple = checkArgs2Pins(args, "pirMotionPin", "ledPin");
+  var magnetPin = tupple.$2;
   var ledPin = tupple.$3;
   switch (tupple.$1) {
     case Hat.nano:
       var hat = NanoHatHub();
       print("Firmeware ${hat.getFirmwareVersion()}");
-      print("$pinInfo: $vibrationPin");
-      print("Led digial pin: $ledPin");
+      print("$pinInfo: $magnetPin");
+      print("Led pin: $ledPin");
 
-      hat.pinMode(vibrationPin, PinMode.input);
+      hat.pinMode(magnetPin, PinMode.input);
       hat.pinMode(ledPin, PinMode.output);
 
-      var old = DigitalValue.high;
+      var old = DigitalValue.low;
       while (true) {
-        var value = hat.digitalRead(vibrationPin);
+        var value = hat.digitalRead(magnetPin);
         print(value);
         if (value != old) {
-          hat.digitalWrite(ledPin, value.invert());
+          hat.digitalWrite(ledPin, value);
         }
         sleep(Duration(milliseconds: wait));
         old = value;
@@ -42,39 +41,43 @@ void main(List<String> args) {
     case Hat.grovePlus:
       var hat = GrovePiPlusHat();
       print("Firmeware ${hat.getFirmwareVersion()}");
-      print("$pinInfo: $vibrationPin");
-      print("Led digial pin: $ledPin");
+      print("$pinInfo: $magnetPin");
+      print("Led pin: $ledPin");
 
-      hat.pinMode(vibrationPin, PinMode.input);
+      hat.pinMode(magnetPin, PinMode.input);
       hat.pinMode(ledPin, PinMode.output);
 
-      var old = DigitalValue.high;
+      var old = DigitalValue.low;
       while (true) {
-        var value = hat.digitalRead(vibrationPin);
+        var value = hat.digitalRead(magnetPin);
         print(value);
         if (value != old) {
-          hat.digitalWrite(ledPin, value.invert());
+          hat.digitalWrite(ledPin, value);
         }
         sleep(Duration(milliseconds: wait));
         old = value;
       }
+
     case Hat.grove:
       var hat = GroveBaseHat();
       print("Firmeware ${hat.getFirmware()}");
       print("Extension hat ${hat.getName()}");
-      print("$pinInfo: $vibrationPin");
-      print("Led digial pin: $ledPin");
-
-      var virbration = GPIO(vibrationPin, GPIOdirection.gpioDirIn);
+      print("$pinInfo: $magnetPin");
+      print("Led pin: $ledPin");
+      var pirMotion = GPIO(magnetPin, GPIOdirection.gpioDirIn);
       var led = GPIO(ledPin, GPIOdirection.gpioDirOut);
       led.write(false);
 
-      var old = true;
+      var old = false;
       while (true) {
-        var value = virbration.read();
-        print(value);
+        var value = pirMotion.read();
         if (value != old) {
-          led.write(!value);
+          led.write(value);
+          if (value == true) {
+            print("Motion detected");
+          } else {
+            print("Watching...");
+          }
         }
         sleep(Duration(milliseconds: wait));
         old = value;